# Store msgrefs in multiple tables to improve queue init performance