--- a/apps/vmq_server/src/vmq_lvldb_store.hrl
+++ b/apps/vmq_server/src/vmq_lvldb_store.hrl
@@ -1,0 +1,2 @@
+-define(TBL_MSG_INIT, vmq_lvldb_init_msg_idx).
+-define(NR_OF_BUCKETS, 12).