Note, the number of tables is pretty arbitrary, just set to the number of msg
store buckets (12) - this could likely be tuned more.

experiment: 25 clients, 200k each, 5m total

current:
msgstore init: 72s
queues init: 43

Ets ops: ~116K/s

new (12 ets tables) phash on subscriberid + persistent term to get table ref:
msgstore init: 54s
queues init: 6s

Ets ops: 833K/s

Improvements:
msgstore init: 133%
queues init: 716% 

Note, I didn't add `{write_concurrency, true}` as this doesn't do anything until OTP 22 (for `ordered_set` tables) and it will use more memory when actually doing something.